backend/pyproject.toml
[project]
name = "pluck-backend"
version = "0.1.0"
description = "Backend for Bird's Nest Inspection System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    
    # Deep Learning
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    
    # Object Storage
    "minio>=7.1.0",
    
    # API
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    
    # Config
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

